backend/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.12
aiofiles>=23.2.1

# Config & validation
pydantic>=2.7.0
pydantic-settings>=2.3.0

# ML core
torch>=2.3.0
torchvision>=0.18.0
numpy>=1.26.0

# Depth estimation (Depth Anything V2)
transformers>=4.44.0
huggingface_hub>=0.24.0
accelerate>=0.33.0

# Image processing
opencv-python-headless>=4.10.0
Pillow>=10.4.0

# Background removal
rembg>=2.0.57
onnxruntime-gpu>=1.18.0

# Super resolution / HD upscaling
basicsr>=1.4.2
realesrgan>=0.3.0

# SVG rasterization
cairosvg>=2.7.1

# HTTP client (for model downloads health checks)
httpx>=0.27.0
